refactor: unify RSI markdown formatting
