+++ rsi_utils.py
@@ -2,11 +2,12 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, Tuple, List
 
 import pandas as pd
 
 from config import shared
+from datetime import datetime
 
 
 def calculate_rsi(prices: pd.Series, period: int = 14) -> Optional[pd.Series]:
@@ -54,3 +55,77 @@ def analyze_extreme_rsi(results: Dict[str, Dict[str, Any]]) -> Dict[str, str]:
                 extreme[f"{symbol} (RSI-6)"] = f"超卖: {r6:.2f}"
 
     return extreme
+
+
+def format_rsi_message(extreme_rsi: Dict[str, str]) -> Tuple[str, str]:
+    """Format extreme RSI readings into a rich Markdown message."""
+    if not extreme_rsi:
+        return "", ""
+
+    overbought_items: List[Tuple[str, str]] = []
+    oversold_items: List[Tuple[str, str]] = []
+
+    for indicator, status in extreme_rsi.items():
+        if "超买" in status:
+            rsi_value = status.split(': ')[1]
+            overbought_items.append((indicator, rsi_value))
+        elif "超卖" in status:
+            rsi_value = status.split(': ')[1]
+            oversold_items.append((indicator, rsi_value))
+
+    title = f"RSI-{len(overbought_items)}个超买,{len(oversold_items)}个超卖信号"
+
+    content_lines: List[str] = [
+        "## 📈 RSI技术指标分析",
+        "",
+        f"🕰️ **检测时间**: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}",
+        "",
+        "---",
+        "",
+    ]
+
+    if overbought_items:
+        content_lines.extend([
+            "### 🔴 **超买区域** `卖出信号`",
+            "",
+            "| 加密货币 | RSI指标 | RSI值 |",
+            "|---------|--------|-------|",
+        ])
+        for indicator, rsi_value in overbought_items:
+            parts = indicator.split(' (')
+            crypto_name = parts[0]
+            rsi_type = parts[1].rstrip(')')
+            content_lines.append(
+                f"| **{crypto_name}** | `{rsi_type}` | **{rsi_value}** |",
+            )
+        content_lines.extend(["", "> 📉 **建议**: 考虑卖出，获利了结", ""])
+
+    if oversold_items:
+        content_lines.extend([
+            "### 🟢 **超卖区域** `买入信号`",
+            "",
+            "| 加密货币 | RSI指标 | RSI值 |",
+            "|---------|--------|-------|",
+        ])
+        for indicator, rsi_value in oversold_items:
+            parts = indicator.split(' (')
+            crypto_name = parts[0]
+            rsi_type = parts[1].rstrip(')')
+            content_lines.append(
+                f"| **{crypto_name}** | `{rsi_type}` | **{rsi_value}** |",
+            )
+        content_lines.extend(["", "> 📈 **建议**: 考虑买入，可能将反弹", ""])
+
+    content_lines.extend([
+        "---",
+        "",
+        "### 📊 RSI指标说明",
+        "",
+        "- **RSI-14**: 14日相对强弱指数 (超买: ≥ 65, 超卖: ≤ 35)",
+        "- **RSI-6**: 6日相对强弱指数 (超买: ≥ 70, 超卖: ≤ 30)",
+        "",
+        "> ⚠️ **免责声明**: 仅供参考，不构成投资建议，请理性投资。",
+    ])
+
+    content = "\n".join(content_lines)
+    return title, content
